All: Rework clockgen-related code
The main goal of this work is to drop the clockgen global instance, as it
uses 135 bytes of global ram space. This instance is only needed when
actively configuring the clock generator, which is a tiny portion of the
time. So allocating so much memory all the time is too expensive.
One annoyance is that the Si5351 class does not seem to expect to be used
in any other way than having Si5351::init called right after construction,
and doing so will entirely reset the chip. There is no obvious way to
instead get the instance to pull the current configuration from the chip.
This means that several attributes will not be properly initialised when
used without calling Si5351::init.
Still, it is possible to use some features without doing so, and especially
it is possible to enable & disable clock outputs.
So:
Declare a new function, getClockGen, which initialises the clock generator
using a possibly stored clock correction factor.
Simplify functions reading from and writing to the correction factor file,
and declare them even when the clock calbration is disabled, so a
previous calibration result keeps being used.
And when simply enabling/disabling clock generator outputs, declare local
instances of the Si5351 class.